drawing_agent/requirements.txt
langchain_community
langchain
langchain-openai
langgraph
langchain-core
omegaconf
hydra-core
python-dotenv
Pillow
opencv-python
numpy
easyocr
torch
pdf2image
poppler-utils
instructor
openai
clearml
pydantic
aiofiles